Dynamic server port (#90)
Make backend and game server port dynamic

When starting the worker backend process and the game server, the application now looks for an available port instead of using a hardcoded one (and not working correctly if this port is already in use). This allows execution on multi-user-systems (terminal servers) and removes the requirement of specific unused ports.

The game server will still be started on port 13050 if available to enable easy connection of manually started clients.