chore(electron): configure auto-updater to not use "v"-prefix
The default setting for the electron-autoupdater when using GitHub is to
check for tags with a prefixed "v" and the version. This can be switched
off by calling setFeedURL[1] with GitHub options hash[2]. Unfortunately, this
leads to repeating the repository name and owner there, which is
otherwise detected automatically from the package.json.

[1] https://www.electron.build/auto-update#module_electron-updater.AppUpdater+setFeedURL
[2] https://www.electron.build/configuration/publish#githuboptions